+++ src/advanced/components/Layout/Header.tsx
@@ -1,7 +1,8 @@
 import { useAtom } from "jotai";
-import { isAdminAtom, toggleAdminAtom, totalCartCountAtom } from "../../store";
+import { isAdminAtom, toggleAdminAtom } from "../../store";
 import { Icons } from "../icons";
 import { SearchBar } from "../searchBar";
+import { totalCartCountAtom } from "../../store/selectors/totalCartCountSelector";
 
 const Header = () => {
   const [isAdmin] = useAtom(isAdminAtom);
